Update
Changed AsyncResource.whenLoaded to call the callback with the
AsyncResource itself, rather than an object of the form {target:
object}.

Added an experiemental sanity check for MDX models. It checks for errors
and warnings in the model data, and reports them.
I hope to add more cases to check.
It can be used with MdxModel.sanityCheck().