Cleaning
Moved the OBJ and BMP handlers into the regular handlers directory, and
added them to the compiler.

The compiler now uses the data in viewer.json to get the needed files.

Added the --unit-tests command-line argument to the compiler.

Added ModelViewer.getRenderedAsBlob(callback)
ModelViewer.getRenderedAsUrl(callback)
ModelViewer.getRenderedAsImage(callback)
Each one will call the callback with the type of object it gets, which
represents what is rendered on the canvas.

Added a  global downloadUrl(url, name) function.

The above functions were removed from UnitTester.

The unit tester now replaces Math.random for the tests, allowing the
tests to be deterministic for particle effects that use randomized
values.

Added loader.js.
This file is used to dynamically load the viewer files.
This should generally not be used outside of testing - for the final
file, it's much preferred if you compile it.
You can see how its used in examples/index.html and tests/index.html.